Extracted two handlers
